2022 - day 8

+++ 2022/day08/test.h
@@ -4,5 +4,5 @@
 
 namespace aoc2022_day08 {
     int part_1(std::string_view path);
-    int part_2(std::string_view path);
+    long long part_2(std::string_view path);
 }
